requirements.txt

mockito==1.0.11
PyHamcrest==1.9.0

